Fixed vendor_id address in libspdm_get_vendor_defined_response Modified libspdm_test_responder_receive_send_rsp_case2 test to use the new API libspdm_register_vendor_callback_func and parameters for VENDOR_DEFINED_RESPONSE Fixed type cast warning in libspdm_vendor_request Fixed type casting warning in libspdm_get_vendor_defined_response

+++ library/spdm_requester_lib/libspdm_req_vendor_request.c
@@ -217,7 +217,7 @@ libspdm_return_t libspdm_vendor_request(void *spdm_context,
         status = libspdm_try_vendor_request(context,
                                             standard_id,
                                             vendor_id_len, (uint8_t *)vendor_id,
-                                            (uint8_t *)request, request_len,
+                                            (uint8_t *)request, (uint16_t)request_len,
                                             (uint8_t *)response, response_len);
         if ((status != LIBSPDM_STATUS_BUSY_PEER) || (retry == 0)) {
             return status;
